dont understand where the error comes in when comparing cpu and gpu for random input values. Currently floating point relative error is about 1e-6 and double error is about 1e-10 (not sure how tight the tolerance should be)
